// packages/lib/src/resources/waf-policy/types.ts
export enum WafPolicyMode {
  Prevention = 'Prevention',
  Detection = 'Detection',
}

/**
 * WAF rule set type.
 */
export enum WafRuleSetType {
  OWASP = 'OWASP',
  Microsoft_BotManagerRuleSet = 'Microsoft_BotManagerRuleSet',
}

/**
 * WAF rule set version.
 */
export enum WafRuleSetVersion {
  V3_2 = '3.2',
  V3_1 = '3.1',
  V3_0 = '3.0',
}

/**
 * WAF state.
 */
export enum WafState {
  Enabled = 'Enabled',
  Disabled = 'Disabled',
}

/**
 * WAF custom rule action.
 */
export enum WafCustomRuleAction {
  Allow = 'Allow',
  Block = 'Block',
  Log = 'Log',
}

/**
 * WAF custom rule type.
 */
export enum WafCustomRuleType {
  MatchRule = 'MatchRule',
  RateLimitRule = 'RateLimitRule',
}

/**
 * WAF match variable.
 */
export enum WafMatchVariable {
  RemoteAddr = 'RemoteAddr',
  RequestMethod = 'RequestMethod',
  QueryString = 'QueryString',
  PostArgs = 'PostArgs',
  RequestUri = 'RequestUri',
  RequestHeaders = 'RequestHeaders',
  RequestBody = 'RequestBody',
  RequestCookies = 'RequestCookies',
}

/**
 * WAF operator.
 */
export enum WafOperator {
  IPMatch = 'IPMatch',
  Equal = 'Equal',
  Contains = 'Contains',
  LessThan = 'LessThan',
  GreaterThan = 'GreaterThan',
  LessThanOrEqual = 'LessThanOrEqual',
  GreaterThanOrEqual = 'GreaterThanOrEqual',
  BeginsWith = 'BeginsWith',
  EndsWith = 'EndsWith',
  Regex = 'Regex',
  GeoMatch = 'GeoMatch',
}